requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
pytz==2023.3
factory-boy==3.3.0
requests==2.31.0
django-cors-headers
